[aoti] Update cshim for all backends
